For issue #272.
Pulls code out from World.cpp and into other files.
Code cleanups towards trimming down World.cpp.
