+++ PlayRho/Dynamics/ContactImpulsesList.cpp
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2017 Louis Langholtz https://github.com/louis-langholtz/PlayRho
+ *
+ * This software is provided 'as-is', without any express or implied
+ * warranty. In no event will the authors be held liable for any damages
+ * arising from the use of this software.
+ *
+ * Permission is granted to anyone to use this software for any purpose,
+ * including commercial applications, and to alter it and redistribute it
+ * freely, subject to the following restrictions:
+ *
+ * 1. The origin of this software must not be misrepresented; you must not
+ *    claim that you wrote the original software. If you use this software
+ *    in a product, an acknowledgment in the product documentation would be
+ *    appreciated but is not required.
+ * 2. Altered source versions must be plainly marked as such, and must not be
+ *    misrepresented as being the original software.
+ * 3. This notice may not be removed or altered from any source distribution.
+ */
+
+#include <PlayRho/Dynamics/ContactImpulsesList.hpp>
+#include <PlayRho/Dynamics/Contacts/VelocityConstraint.hpp>
+
+namespace playrho {
+namespace d2 {
+
+/// @brief Gets the contact impulses for the given velocity constraint.
+ContactImpulsesList GetContactImpulses(const VelocityConstraint& vc)
+{
+    ContactImpulsesList impulse;
+    const auto count = vc.GetPointCount();
+    for (auto j = decltype(count){0}; j < count; ++j)
+    {
+        impulse.AddEntry(GetNormalImpulseAtPoint(vc, j), GetTangentImpulseAtPoint(vc, j));
+    }
+    return impulse;
+}
+
+} // namespace d2
+} // namespace playrho
